# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Cosmosdb sample
  description: Cosmosdb sample
  version: 1.0.0
#servers:
#  - url: test
paths:
  /users:
    get:
      operationId: findUsers
      responses:
        '200':
          description: User response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      operationId: "createUser"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        '200':
          description:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{id}:
    get:
      summary: "Get user by user id"
      description: ""
      operationId: "getUserById"
      parameters:
        - name: id
          in: path
          description: ID of pet to fetch
          required: true
          schema:
            type: string
      responses:
        '200':
          description: pet response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: "Invalid id supplied"
        "404":
          description: "User not found with given id"
components:
  schemas:
    Address:
      type: "object"
      required:
        - street
        - postalcode
        - city
      properties:
        street:
          type: "string"
        postalcode:
          type: "string"
        city:
          type: "string"
      xml:
        name: "Address"
    User:
      type: "object"
      required:
        - id
        - firstName
        - lastName
      properties:
        id:
          type: string
        firstName:
          type: "string"
        lastName:
          type: "string"
        addresses:
          type: "array"
          xml:
            name: "Address"
            wrapped: true
          items:
            $ref: "#/components/schemas/Address"

    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string